Implement Entities extensions code generation
Added a Tools panel to the Editor's top menu, enabling the generation of extension methods for all Entity components. To include a component in the code generation, it must be registered in the _entityValuesToType dictionary of the EntityExtensionsGenerator class.
Also added the EntityExtensions class for writing Entity extensions manually in the future.